TASK: cleanup and Refactoring
